refactor(io.output): :recycle: move the Pager class to a dedicated module
