+++ rocrate_validator/cli/utils.py
@@ -13,19 +13,11 @@
 # limitations under the License.
 
 import os
-import pydoc
 import re
 import textwrap
-from typing import Any, Optional
+from typing import Optional
 
-from rich.console import Console as BaseConsole
-from rich.padding import Padding
-from rich.pager import Pager
-from rich.rule import Rule
-from rich.text import Text
-
-from rocrate_validator import log as logging
-from rocrate_validator.utils import get_version
+import rocrate_validator.log as logging
 
 # set up logging
 logger = logging.getLogger(__name__)
@@ -52,17 +44,6 @@ def format_text(text: str,
 def get_app_header_rule() -> Text:
     return Padding(Rule(f"\n[bold][cyan]ROCrate Validator[/cyan] (ver. [magenta]{get_version()}[/magenta])[/bold]",
                         style="bold cyan"), (1, 2))
-
-
-class SystemPager(Pager):
-    """Uses the pager installed on the system."""
-
-    def _pager(self, content: str) -> Any:
-        return pydoc.pipepager(content, "less -R -K")
-
-    def show(self, content: str) -> None:
-        """Use the same pager used by pydoc."""
-        self._pager(content)
 
 
 class Console(BaseConsole):
